0.3.5
修改表的结构修改，将原项目方法表分解成项目表、类型方法表、点位方法表；
对应修改了任务单保存、载入；方法表的保存、载入；
修改了数据传输格式，数据长度由16位记录改为32位记录，以增加对长数据的支持；
服务端的数据库处理优化，修改对多条sql语句的处理为分开处理，以便正确识别错误。
优化了客户端任务单载入和保存的逻辑，使用合并操作功能减少网络来回传输。
完善服务器日志功能。完善了连接池功能。

+++ depends/CELLMsgStream.cpp
@@ -20,20 +20,22 @@ CELLWriteStream::CELLWriteStream(char *pData, int nSize, bool bDelete)
     :CELLStream(pData, nSize, bDelete)
 {
     //预先占领消息长度所需空间
-    Write<uint16_t>(0);
+    //    Write<uint16_t>(0);
+    Write<uint32_t>(0);
 }
 
 CELLWriteStream::CELLWriteStream(int nSize)
     :CELLStream(nSize)
 {
     //预先占领消息长度所需空间
-    Write<uint16_t>(0);
+//    Write<uint16_t>(0);
+    Write<uint32_t>(0);
 }
 
 void CELLWriteStream::finsh()
 {
     int pos = length();
     setWritePos(0);
-    Write<uint16_t>(pos);
+    Write<uint32_t>(pos);
     setWritePos(pos);
 }
